Security updates : 4pm : 13 may

+++ src/main/java/com/example/final_project/Service/ItemSaleService.java
@@ -5,6 +5,7 @@ import com.example.final_project.Model.Accountant;
 import com.example.final_project.Model.ItemSale;
 import com.example.final_project.Model.Product;
 import com.example.final_project.Model.Sales;
+import com.example.final_project.Repository.AccountantRepository;
 import com.example.final_project.Repository.ItemSaleRepository;
 import com.example.final_project.Repository.ProductRepository;
 import com.example.final_project.Repository.SalesRepository;
@@ -24,6 +25,9 @@ public class ItemSaleService {
     private final ItemSaleRepository itemSaleRepository;
     private final ProductRepository productRepository;
     private final SalesRepository salesRepository;
+    private final SalesService salesService;
+    private final AccountantRepository accountantRepository;
+
 
     public List<ItemSale> getAllItemSales() {
         return itemSaleRepository.findAll();
@@ -43,13 +47,35 @@ public class ItemSaleService {
         itemSaleRepository.save(itemSale);
     }
 
+    public void removeItemFromSale(Integer accounterId, Integer itemId,Integer saleId) {
 
-    public void deleteItemSale(Integer id) {
-        ItemSale itemSale = itemSaleRepository.findItemSaleById(id);
-                if(itemSale==null){
-                  throw new ApiException("ItemSale not found");
-                }
+        Accountant accountant = accountantRepository.findAccountantById(accounterId);
+        if (accountant == null) {
+            throw new ApiException("Accountant not found");
+        }
+        ItemSale itemSale = itemSaleRepository.findItemSaleById(itemId);
+        if (itemSale == null) {
+            throw new ApiException("ItemSale not found");
+        }
+        Sales currentSale = salesRepository.findSalesById(saleId);
+        if (currentSale == null) {
+            throw new ApiException("Sale not found");
+        }
         itemSaleRepository.delete(itemSale);
+
+        Product product1 = itemSale.getProduct();
+
+        if (currentSale.getSalesStatus().equals("CONFIRMED")) {
+            throw new ApiException("Cannot remove item from a confirmed invoice");
+        }
+
+
+        product1.setStock(product1.getStock()+itemSale.getQuantity());
+        salesService.updateCalculations(saleId);
+        productRepository.save(product1);
+        itemSaleRepository.delete(itemSale);
+        salesRepository.save(currentSale);
+
     }
 
     public List<ItemSale> getItemSalesBySalesId(Integer salesId) {
